+++ src/Service/CloudFormation/src/Enum/StackDriftDetectionStatus.php
@@ -6,7 +6,7 @@ namespace AsyncAws\CloudFormation\Enum;
  * The status of the stack drift detection operation.
  *
  * - `DETECTION_COMPLETE`: The stack drift detection operation has successfully completed for all resources in the stack
- *   that support drift detection. (Resources that do not currently support stack detection remain unchecked.)
+ *   that support drift detection. (Resources that don't currently support stack detection remain unchecked.)
  *   If you specified logical resource IDs for CloudFormation to use as a filter for the stack drift detection
  *   operation, only the resources with those logical IDs are checked for drift.
  * - `DETECTION_FAILED`: The stack drift detection operation has failed for at least one resource in the stack. Results